+++ src/Utils/PriceTypes.php
@@ -1,0 +1,65 @@
+<?php
+
+/**
+ * TechDivision\Import\Product\Bundle\Utils\PriceTypes
+ *
+ * NOTICE OF LICENSE
+ *
+ * This source file is subject to the Open Software License (OSL 3.0)
+ * that is available through the world-wide-web at this URL:
+ * http://opensource.org/licenses/osl-3.0.php
+ *
+ * PHP version 5
+ *
+ * @author    Tim Wagner <t.wagner@techdivision.com>
+ * @copyright 2019 TechDivision GmbH <info@techdivision.com>
+ * @license   http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
+ * @link      https://github.com/techdivision/import-product-bundle
+ * @link      http://www.techdivision.com
+ */
+
+namespace TechDivision\Import\Product\Bundle\Utils;
+
+/**
+ * Utility class containing the available bundle price types.
+ *
+ * @author    Tim Wagner <t.wagner@techdivision.com>
+ * @copyright 2019 TechDivision GmbH <info@techdivision.com>
+ * @license   http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
+ * @link      https://github.com/techdivision/import-product-bundle
+ * @link      http://www.techdivision.com
+ */
+class PriceTypes
+{
+
+    /**
+     * The value for the price type 'fixed'.
+     *
+     * @var integer
+     */
+    const FIXED = 0;
+
+    /**
+     * The value for the price type 'percent'.
+     *
+     * @var integer
+     */
+    const PERCENT = 1;
+
+    /**
+     * This is a utility class, so protect it against direct
+     * instantiation.
+     */
+    private function __construct()
+    {
+    }
+
+    /**
+     * This is a utility class, so protect it against cloning.
+     *
+     * @return void
+     */
+    private function __clone()
+    {
+    }
+}